Small cleanups

+++ old/reduction.c
@@ -3,7 +3,7 @@
 #include <pthread.h>
 
 /* Number of threads */
-int P;
+unsigned int P;
 
 #define PTHREAD_CREATE(a, b, c, d)                  \
     {                                               \
@@ -16,7 +16,7 @@ int P;
 typedef struct {
     double *array1d;
     double *array2d;
-    int tid;
+    unsigned int tid;
     size_t size;
 } InitArg;
 
@@ -102,7 +102,7 @@ int main(int argc, char **argv)
     double *arr1d = ShrayMalloc(n * n, n * n * sizeof(double));
     double *arr2d = ShrayMalloc(n, n * n * sizeof(double));
 
-    for (int tid = 0; tid < P; tid++) {
+    for (unsigned int tid = 0; tid < P; tid++) {
         inits[tid].array1d = arr1d;
         inits[tid].array2d = arr2d;
         inits[tid].tid = tid;
@@ -110,7 +110,7 @@ int main(int argc, char **argv)
         PTHREAD_CREATE(&threads[tid], NULL, parallelInit, &inits[tid]);
     }
 
-    for (size_t tid = 0; tid < P; tid++) {
+    for (unsigned int tid = 0; tid < P; tid++) {
         pthread_join(threads[tid], returnValues[tid]);
     }
 
@@ -122,15 +122,15 @@ int main(int argc, char **argv)
     double sum1 = 0.0;
     double sum2 = 0.0;
 
-    for (int tid = 0; tid < P; tid++) {
+    for (unsigned int tid = 0; tid < P; tid++) {
         PTHREAD_CREATE(&threads[tid], NULL, parallelReduce, &inits[tid]);
     }
 
-    for (size_t tid = 0; tid < P; tid++) {
+    for (unsigned int tid = 0; tid < P; tid++) {
         pthread_join(threads[tid], returnSums[tid]);
     }
 
-    for (size_t tid = 0; tid < P; tid++) {
+    for (unsigned int tid = 0; tid < P; tid++) {
         sum1 += ((double *)returnSums[tid])[0];
         sum2 += ((double *)returnSums[tid])[1];
     }
